+++ tests/test_xml.py
@@ -54,12 +54,12 @@ def test_parse_xml_tracks_v5():
 
     assert xml.num_tracks == 6
 
-    track = xml.get_track(1)
+    track = xml.get_track(0)
     assert track.Name == "NOISE"
     assert len(track.tempos) == 0
     assert len(track.marks) == 0
 
-    track = xml.get_track(5)
+    track = xml.get_track(4)
 
     tempo = track.tempos[0]
     assert tempo.Bpm == 128.0
@@ -73,3 +73,18 @@ def test_parse_xml_tracks_v5():
     assert positions[0].Type == "cue"
     assert positions[0].Start == 0.025
     assert positions[0].Num == -1
+
+
+def test_update_track_count():
+    xml = RekordboxXml()
+    track1 = xml.add_track("C:/path/to/file1.wav")
+    assert xml.num_tracks == 1
+
+    track2 = xml.add_track("C:/path/to/file2.wav")
+    assert xml.num_tracks == 2
+
+    xml.remove_track(track1)
+    assert xml.num_tracks == 1
+
+    xml.remove_track(track2)
+    assert xml.num_tracks == 0